Add files via upload
PROBLEM STATEMENT

The IIT Jodhpur mess hall is in need of a better queue management system during peak times, such as on Republic Day. On this day, the length of the queue became so large that the mess services decided to split the queue into two separate lines. To accomplish this, the queue will be split in half, with the people from first to the middle being asked to form a new line, but this time moving in the opposite direction.

You're given an array of IDs of the students. Your task is to design and implement a program using queue and stack data structures to assist the mess services in efficiently splitting the queue and managing the lines in an orderly manner. Print the IDs of the students from left to right as the output.


NOTE: You have to use QUEUE and STACK to solve this problem. Create queue and stack utility functions.

INPUT FORMAT
The first line contains N, the number of students.
The second line contains  N-spaced integers denoting the IDs of the students.

OUTPUT FORMAT
Print spaced integers (in the same order as input) denoting the IDs of the students after the split. 

CONSTRAINTS
1 ≤ N ≤ 10^4
-10^5 ≤ A[i] ≤ 10^5

SAMPLE INPUT 1
5
3 4 6 5 7

SAMPLE OUTPUT 1
4 3 6 5 7

EXPLANATION
So, here 5 people need to be split in 2 lines. The split is made from index 2 and the students before index 2 are sent to another line in opposite direction. And now the traversal is from the end of one line to the end of other line.

SAMPLE INPUT 2
4
5 22 32 14

SAMPLE OUTPUT 2
22 5 32 14

SAMPLE INPUT 3
1
5

SAMPLE OUTPUT 3
5